Add empty jetty-base/tmp/ directory.
We set the location of Jetty's temp directory via the
org.eclipse.jetty.webapp.basetempdir attribute of the IdP WebAppContext
in idp.xml, and this directory must exist. If it does not, Jetty uses
the system default temporary directory, which will not be cleaned up by
the integration test harness following successful tests.